Refactored ArgumentParser
