+++ CMakeLists.txt
@@ -15,7 +15,10 @@ endif ()
 # Building
 project(KI VERSION 0.0.1 DESCRIPTION "KI for SoPra")
 
-set(SOURCES )
+set(SOURCES
+        ${CMAKE_SOURCE_DIR}/src/Util/Logging.cpp
+        ${CMAKE_SOURCE_DIR}/src/Util/Timer.cpp
+        ${CMAKE_SOURCE_DIR}/src/Util/ArgumentParser.cpp)
 
 set(LIBS pthread stdc++fs SopraGameLogic SopraMessages SopraNetwork)
